Rename "MessageListEntry" component to "ConversationListEntry"
