+++ src/components/ConversationListEntry.svelte
